+++ test/plugin/store_dimensions_test.rb
@@ -21,7 +21,9 @@ describe Shrine::Plugins::StoreDimensions do
     end
 
     it "returns nil for non-images" do
-      assert_nil @shrine.extract_dimensions(fakeio)
+      assert_logged /SHRINE WARNING: Error occurred/ do
+        assert_nil @shrine.extract_dimensions(fakeio)
+      end
     end
   end
 
@@ -40,7 +42,9 @@ describe Shrine::Plugins::StoreDimensions do
     end
 
     it "returns nil for non-images" do
-      assert_nil @shrine.extract_dimensions(fakeio)
+      assert_logged /SHRINE WARNING: Error occurred/ do
+        assert_nil @shrine.extract_dimensions(fakeio)
+      end
     end
   end unless ENV["CI"]
 
@@ -59,7 +63,9 @@ describe Shrine::Plugins::StoreDimensions do
     end
 
     it "returns nil for non-images" do
-      assert_nil @shrine.extract_dimensions(fakeio)
+      assert_logged /SHRINE WARNING: Error occurred/ do
+        assert_nil @shrine.extract_dimensions(fakeio)
+      end
     end
   end unless ENV["CI"]